requirements.txt
# Core API
fastapi
uvicorn
python-dotenv
requests

# The "Brain" Logic (RAG & AI)
langchain
langchain-community
langchain-postgres
langchain-huggingface
psycopg[binary]
pgvector
sentence-transformers

# Data Handling
pandas
numpy

# The "Mouth" (Ollama Client)
ollama

# PDF Parsing (CPU based)
pymupdf